Lock player when a View is open + fix text color in Button

+++ src/interactable.cpp
@@ -149,7 +149,7 @@ void Interactable::interact() {
 			std::cout << error.what() << std::endl;
 		}
 		if (result.status() != sol::call_status::ok) {
-			printf("uh oh: %i \n", result.status());
+			printf("uh oh: %i \n", static_cast<int>(result.status()));
 		}
 
 		if (state["interact"].valid()) {
